Split admin controller into many
