+++ app/views/admin/home/_latest_shared_seasons.html.erb
