+++ app/views/admin/home/_seasons.html.erb
