+++ app/views/admin/home/_users.html.erb
